\name{add_heatmap-dispatch}
\alias{add_heatmap}
\title{
Method dispatch page for add_heatmap
}
\description{
Method dispatch page for \code{add_heatmap}.
}
\section{Dispatch}{
\code{add_heatmap} can be dispatched on following classes:

\itemize{
\item \code{\link{add_heatmap,HeatmapAnnotation-method}}, \code{\link{HeatmapAnnotation-class}} class method
\item \code{\link{add_heatmap,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{add_heatmap,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
