\name{add_heatmap-Heatmap-method}
\alias{add_heatmap,Heatmap-method}
\title{
add two heatmaps as a heatmap list  


}
\description{
add two heatmaps as a heatmap list  


}
\usage{
\S4method{add_heatmap}{Heatmap}(object, ht)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object}
  \item{ht}{a \code{\link{Heatmap}} object}

}
\value{
a \code{\link{HeatmapList}} object  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
