\name{add_heatmap-HeatmapAnnotation-method}
\alias{add_heatmap,HeatmapAnnotation-method}
\title{
Add two row annotations or add heatmaps as a heatmap list  


}
\description{
Add two row annotations or add heatmaps as a heatmap list  


}
\usage{
\S4method{add_heatmap}{HeatmapAnnotation}(object, x)
}
\arguments{

  \item{object}{a \code{\link{HeatmapAnnotation}} object.}
  \item{x}{a \code{\link{Heatmap}} object, a \code{\link{HeatmapAnnotation}} object or a \code{\link{HeatmapList}} object.}

}
\details{
There is a shortcut function \code{+.HeatmapAnnotation}.  


}
\value{
A \code{\link{HeatmapList}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
