\name{AdditiveUnit-class}
\docType{class}
\alias{AdditiveUnit-class}
\alias{AdditiveUnit}
\title{
A conceptual class  


}
\description{
A conceptual class  


}
\details{
This class is a super class for \code{\link{Heatmap}}, \code{\link{HeatmapList}} and \code{\link{HeatmapAnnotation}} classes.  


}
