\name{AdditiveUnit}
\alias{AdditiveUnit}
\title{
Constructor Method for AdditiveUnit Class
}
\description{
Constructor Method for AdditiveUnit Class
}
\usage{
AdditiveUnit(...)
}
\arguments{

  \item{...}{Black hole arguments.}

}
\details{
This method is not used in the package.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
