\name{anno_block}
\alias{anno_block}
\title{
Block annotation
}
\description{
Block annotation
}
\usage{
anno_block(gp = gpar(), labels = NULL, labels_gp = gpar(), labels_rot = ifelse(which == "row", 90, 0),
    which = c("column", "row"), width = NULL, height = NULL)
}
\arguments{

  \item{gp}{Graphic parameters}
  \item{labels}{-labels}
  \item{labels_gp}{-labels_gp}
  \item{labels_rot}{-labels_rot}
  \item{which}{-which}
  \item{width}{-width}
  \item{height}{-height}

}
\examples{
# There is no example
NULL

}
