\name{anno_boxplot}
\alias{anno_boxplot}
\title{
Using boxplot as annotation  


}
\description{
Using boxplot as annotation  


}
\usage{
anno_boxplot(x, which = c("column", "row"), gp = gpar(),
    pch = 16, size = unit(2, "mm"))
}
\arguments{

  \item{x}{a matrix or a list}
  \item{which}{column annotation or row annotation}
  \item{gp}{graphic parameters}
  \item{pch}{point type}
  \item{size}{point size}

}
\value{
A function  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
