\name{anno_link}
\alias{anno_link}
\title{
Label Markers Annotation
}
\description{
Label Markers Annotation
}
\usage{
anno_link(...)
}
\arguments{

  \item{...}{Pass to \code{\link{anno_mark}}.}

}
\details{
\code{\link{anno_link}} is deprecated, please use \code{\link{anno_mark}} instead.
}
\examples{
# There is no example
NULL

}
