\name{anno_points}
\alias{anno_points}
\title{
Using points as annotation  


}
\description{
Using points as annotation  


}
\usage{
anno_points(x, which = c("column", "row"), gp = gpar(), pch = 16,
    size = unit(2, "mm"))
}
\arguments{

  \item{x}{a vector}
  \item{which}{column annotation or row annotation}
  \item{gp}{graphic parameters}
  \item{pch}{point type}
  \item{size}{point size}

}
\value{
A function  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
