\name{annotation_axis_grob}
\alias{annotation_axis_grob}
\title{
Grob for annotation axis
}
\description{
Grob for annotation axis
}
\usage{
annotation_axis_grob(at = NULL, labels = at, labels_rot = 0, gp = gpar(),
    side = "left", facing = "outside")
}
\arguments{

  \item{at}{Break values.}
  \item{labels}{Corresponding labels.}
  \item{labels_rot}{Rotations of labels.}
  \item{gp}{Graphic parameters.}
  \item{side}{side of the axis to the annotation viewport.}
  \item{facing}{Facing of the axis.}

}
\value{
A \code{\link[grid]{grob}} object.
}
\examples{
# There is no example
NULL
}
