\name{ColorMapping-class}
\docType{class}
\alias{ColorMapping-class}
\alias{ColorMapping}
\title{
Class to map values to colors  


}
\description{
Class to map values to colors  


}
\details{
The \code{\link{ColorMapping}} class handles color mapping with both discrete values and continuous values.  


}
\section{Methods}{
The \code{\link{ColorMapping}} class provides following methods:  

\itemize{
  \item \code{\link{initialize,ColorMapping-method}}: contructor methods
  \item \code{\link{map,ColorMapping-method}}: mapping values to colors
  \item \code{\link{color_mapping_legend,ColorMapping-method}}: draw legend
}


}
\section{Methods}{
The \code{\link{ColorMapping}} class provides following methods:  

\itemize{
  \item \code{\link{initialize,ColorMapping-method}}: contructor methods
  \item \code{\link{map,ColorMapping-method}}: mapping values to colors
  \item \code{\link{color_mapping_legend,ColorMapping-method}}: draw legend
}
}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
