\name{column_anno_barplot}
\alias{column_anno_barplot}
\title{
Column annotation which is represented as barplots
}
\description{
Column annotation which is represented as barplots
}
\usage{
column_anno_barplot(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_barplot}}}

}
\details{
A wrapper of \code{\link{anno_barplot}} with pre-defined \code{which} to \code{column}.
}
