\name{column_anno_boxplot}
\alias{column_anno_boxplot}
\title{
Column annotation which is represented as boxplots
}
\description{
Column annotation which is represented as boxplots
}
\usage{
column_anno_boxplot(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_boxplot}}}

}
\details{
A wrapper of \code{\link{anno_boxplot}} with pre-defined \code{which} to \code{column}.
}
\value{
See help page of \code{\link{anno_boxplot}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
