\name{column_anno_density}
\alias{column_anno_density}
\title{
Column annotation which is represented as density plot
}
\description{
Column annotation which is represented as density plot
}
\usage{
column_anno_density(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_density}}}

}
\details{
A wrapper of \code{\link{anno_density}} with pre-defined \code{which} to \code{column}.
}
\value{
See help page of \code{\link{anno_density}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
