\name{column_anno_histogram}
\alias{column_anno_histogram}
\title{
Column annotation which is represented as histogram
}
\description{
Column annotation which is represented as histogram
}
\usage{
column_anno_histogram(...)}
\arguments{

  \item{...}{pass to \code{\link{anno_histogram}}}
}
\details{
A wrapper of \code{\link{anno_histogram}} with pre-defined \code{which} to \code{column}.
}
