\name{column_anno_points}
\alias{column_anno_points}
\title{
Column annotation which is represented as points
}
\description{
Column annotation which is represented as points
}
\usage{
column_anno_points(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_points}}}

}
\details{
A wrapper of \code{\link{anno_points}} with pre-defined \code{which} to \code{column}.
}
\value{
See help page of \code{\link{anno_points}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
