\name{column_anno_text}
\alias{column_anno_text}
\title{
Column annotation which is represented as text
}
\description{
Column annotation which is represented as text
}
\usage{
column_anno_text(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_text}}}

}
\details{
A wrapper of \code{\link{anno_text}} with pre-defined \code{which} to \code{column}.
}
