\name{column_dend-dispatch}
\alias{column_dend}
\title{
Method dispatch page for column_dend
}
\description{
Method dispatch page for \code{column_dend}.
}
\section{Dispatch}{
\code{column_dend} can be dispatched on following classes:

\itemize{
\item \code{\link{column_dend,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{column_dend,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
