\name{column_order-dispatch}
\alias{column_order}
\title{
Method dispatch page for column_order
}
\description{
Method dispatch page for \code{column_order}.
}
\section{Dispatch}{
\code{column_order} can be dispatched on following classes:

\itemize{
\item \code{\link{column_order,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{column_order,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
