\name{columnAnnotation}
\alias{columnAnnotation}
\title{
Construct Column Annotations
}
\description{
Construct Column Annotations
}
\usage{
columnAnnotation(...)
}
\arguments{

  \item{...}{Pass to \code{\link{HeatmapAnnotation}}.}

}
\details{
The function is identical to

  \preformatted{
    HeatmapAnnotation(..., which = "column")  }
}
\value{
A \code{\link{HeatmapAnnotation-class}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
