\name{complement_size}
\alias{complement_size}
\title{
Complement Set Size
}
\description{
Complement Set Size
}
\usage{
complement_size(m)
}
\arguments{

  \item{m}{A combination matrix returned by \code{\link{make_comb_mat}}.}

}
\value{
If there is no complement set, it returns zero.
}
\examples{
# There is no example
NULL

}
