\name{component_height-Heatmap-method}
\alias{component_height,Heatmap-method}
\title{
height of each heatmap component  


}
\description{
height of each heatmap component  


}
\usage{
\S4method{component_height}{Heatmap}(object, k = 1:9)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object}
  \item{k}{which components}

}
\value{
a \code{\link[grid]{unit}} object  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
