\name{component_height-HeatmapList-method}
\alias{component_height,HeatmapList-method}
\title{
height of components  


}
\description{
height of components  


}
\usage{
\S4method{component_height}{HeatmapList}(object, k = 1:7)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{k}{components}

}
