\name{component_width-dispatch}
\alias{component_width}
\title{
Method dispatch page for component_width
}
\description{
Method dispatch page for \code{component_width}.
}
\section{Dispatch}{
\code{component_width} can be dispatched on following classes:

\itemize{
\item \code{\link{component_width,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{component_width,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
