\name{component_width-Heatmap-method}
\alias{component_width,Heatmap-method}
\title{
width of each heatmap component  


}
\description{
width of each heatmap component  


}
\usage{
\S4method{component_width}{Heatmap}(object, k = 1:7)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object}
  \item{k}{which components}

}
\value{
a \code{\link[grid]{unit}} object  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
