\name{component_width-HeatmapList-method}
\alias{component_width,HeatmapList-method}
\title{
width of each components  


}
\description{
width of each components  


}
\usage{
\S4method{component_width}{HeatmapList}(object, k = 1:7)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{k}{components}

}
