\name{copy_all-dispatch}
\alias{copy_all}
\title{
Method dispatch page for copy_all
}
\description{
Method dispatch page for \code{copy_all}.
}
\section{Dispatch}{
\code{copy_all} can be dispatched on following classes:

\itemize{
\item \code{\link{copy_all,AnnotationFunction-method}}, \code{\link{AnnotationFunction-class}} class method
\item \code{\link{copy_all,SingleAnnotation-method}}, \code{\link{SingleAnnotation-class}} class method
}
}
\examples{
# no example
NULL


}
