\name{decorate_column_dend}
\alias{decorate_column_dend}
\title{
Decorate Heatmap Column Dendrograms
}
\description{
Decorate Heatmap Column Dendrograms
}
\usage{
decorate_column_dend(..., envir = new.env(parent = parent.frame()))
}
\arguments{

  \item{...}{Pass to \code{\link{decorate_dend}}.}
  \item{envir}{Where to look for variables inside \code{code}.}

}
\details{
This is a wrapper function which pre-defined \code{which} argument in \code{\link{decorate_dend}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
