\name{decorate_column_hclust}
\alias{decorate_column_hclust}
\title{
Decorate heatmap dendrogram on columns
}
\description{
Decorate heatmap dendrogram on columns
}
\usage{
decorate_column_hclust(...)
}
\arguments{

  \item{...}{pass to \code{\link{decorate_hclust}}}

}
\details{
This is a wrapper function which pre-defined \code{which} argument in \code{\link{decorate_hclust}}.
}
