\name{decorate_row_hclust}
\alias{decorate_row_hclust}
\title{
Decorate heatmap dendrogram on rows
}
\description{
Decorate heatmap dendrogram on rows
}
\usage{
decorate_row_hclust(...)
}
\arguments{

  \item{...}{pass to \code{\link{decorate_hclust}}}

}
\details{
This is a wrapper function which pre-defined \code{which} argument in \code{\link{decorate_hclust}}.
}
