\name{decorate_row_names}
\alias{decorate_row_names}
\title{
Decorate heatmap row names
}
\description{
Decorate heatmap row names
}
\usage{
decorate_row_names(...)
}
\arguments{

  \item{...}{pass to \code{\link{decorate_dimnames}}}

}
\details{
This is a wrapper function which pre-defined \code{which} argument in \code{\link{decorate_dimnames}}.
}
