\name{decorate_row_title}
\alias{decorate_row_title}
\title{
Decorate heatmap row title
}
\description{
Decorate heatmap row title
}
\usage{
decorate_row_title(...)
}
\arguments{

  \item{...}{pass to \code{\link{decorate_title}}}

}
\details{
This is a wrapper function which pre-defined \code{which} argument in \code{\link{decorate_title}}.
}
