\name{default_get_type}
\alias{default_get_type}
\title{
Default get_type for oncoPrint()
}
\description{
Default get_type for oncoPrint()
}
\usage{
default_get_type(x)
}
\arguments{

  \item{x}{A strings which encode multiple altertations.}

}
\details{
It recognizes following separators: \code{;:,|}.
}
\examples{
# There is no example
NULL

}
