\name{default_upset_right_annotation}
\alias{default_upset_right_annotation}
\title{
Default UpSet Right Annotation
}
\description{
Default UpSet Right Annotation
}
\usage{
default_upset_right_annotation(m)
}
\arguments{

  \item{m}{A combination matrix which is as same as the one for \code{\link{UpSet}}.}

}
\details{
The default right annotation is actually barplot implemented by \code{\link{anno_barplot}}. For
how to set the right annotation or left annotation in \code{\link{UpSet}}, please refer to \code{\link{UpSet}}.
}
\examples{
# There is no example
NULL

}
