\name{default_upset_top_annotation}
\alias{default_upset_top_annotation}
\title{
Default UpSet Top Annotation
}
\description{
Default UpSet Top Annotation
}
\usage{
default_upset_top_annotation(m)
}
\arguments{

  \item{m}{A combination matrix which is as same as the one for \code{\link{UpSet}}.}

}
\details{
The default top annotation is actually barplot implemented by \code{\link{anno_barplot}}. For
how to set the top annotation or bottom annotation in \code{\link{UpSet}}, please refer to \code{\link{UpSet}}.
}
\examples{
# There is no example
NULL

}
