% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.dendrogram.R
\name{dend_heights}
\alias{dend_heights}
\title{Height of the Dendrograms}
\usage{
dend_heights(x)
}
\arguments{
\item{x}{a \code{\link{dendrogram}} object or a list of
\code{\link{dendrogram}} objects.}
}
\description{
Height of the Dendrograms
}
\examples{


# There is no example
NULL



}
