% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-utils.R
\name{dim.Heatmap}
\alias{dim.Heatmap}
\title{Dimension of the Heatmap}
\usage{
\method{dim}{Heatmap}(x)
}
\arguments{
\item{x}{A \code{\link{Heatmap-class}} object.}
}
\description{
Dimension of the Heatmap
}
\examples{


# There is no example
NULL



}
