\name{draw-SingleAnnotation-method}
\alias{draw,SingleAnnotation-method}
\title{
Draw the single annotation  


}
\description{
Draw the single annotation  


}
\usage{
\S4method{draw}{SingleAnnotation}(object, index)
}
\arguments{

  \item{object}{a \code{\link{SingleAnnotation}} object.}
  \item{index}{the index}

}
\details{
A viewport is created.  


}
\value{
No value is returned.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
