\name{draw_heatmap_legend-HeatmapList-method}
\alias{draw_heatmap_legend,HeatmapList-method}
\title{
draw heatmap legend  


}
\description{
draw heatmap legend  


}
\usage{
\S4method{draw_heatmap_legend}{HeatmapList}(object, side = c("right", "left", "top", "bottom"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{side}{side}

}
