\name{draw_heatmap_list-HeatmapList-method}
\alias{draw_heatmap_list,HeatmapList-method}
\alias{draw_heatmap_list}
\title{
plot list of heatmaps  


}
\description{
plot list of heatmaps  


}
\usage{
\S4method{draw_heatmap_list}{HeatmapList}(object, hgap = unit(2, "mm"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{hgap}{gap}

}
