\name{draw_title-HeatmapList-method}
\alias{draw_title,HeatmapList-method}
\title{
draw title  


}
\description{
draw title  


}
\usage{
\S4method{draw_title}{HeatmapList}(object, title, which = c("row", "column"),
    side = ifelse(which == "row", "right", "bottom"), gp = NULL)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{title}{title}
  \item{which}{which}
  \item{side}{side}
  \item{gp}{graphic parameters for drawing text}

}
