\name{enhanced_basicplot}
\alias{enhanced_basicplot}
\title{
Enhanced version of basic barplot and boxplot
}
\description{
Enhanced version of basic barplot and boxplot
}
\usage{
enhanced_basicplot(data, ..., ylim = NULL,
    ylab = deparse(substitute(data)), title = NULL, title_gp = gpar(fontsize = 14),
    type = c("boxplot", "barplot"), width = 0.8, gp = gpar(),
    pch = 1, size = unit(2, "mm"), axis_gp = gpar(fontsize = 8),
    padding = unit(c(2, 18, 2, 2), "mm"),
    heatmap_legend_list = list())
}
\arguments{

  \item{data}{a matrix, a list or a simple numeric vector. If your data is a data frame please convert it to a matrix in the first place.}
  \item{...}{pass to \code{\link{Heatmap}}}
  \item{ylim}{ranges on y axis}
  \item{ylab}{label on y axis}
  \item{title}{title of the plot}
  \item{title_gp}{graphic parameters for the title}
  \item{type}{type of the plot}
  \item{width}{relative width of the bar or box}
  \item{gp}{graphic parameters for hte bar or box}
  \item{pch}{shape of outlier points in the boxplot}
  \item{size}{size of hte outlier points in the boxplot}
  \item{axis_gp}{graphic parameters for the axis}
  \item{padding}{padding of the plot}
  \item{heatmap_legend_list}{a list of \code{\link[grid]{grob}} which contains legend. It can be generated by \code{\link{color_mapping_legend,ColorMapping-method}}.}

}
\details{
This function adds annotations to the barplot or boxplot.

This function is still quite experimental.
}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
mat = matrix(runif(100), 10)
enhanced_basicplot(mat)
ha = HeatmapAnnotation(char = sample(letters[1:2], 10, replace = TRUE),
                       num = runif(10))
enhanced_basicplot(mat, top_annotation = ha)
enhanced_basicplot(mat, type = "barplot", top_annotation = ha)
}
