\name{[.gridtext}
\alias{Extract.gridtext}
\alias{[.gridtext}
\title{
Subset method of gridtext class
}
\description{
Subset method of gridtext class
}
\usage{
\method{[}{gridtext}(x, index)
}
\arguments{

  \item{x}{A vector of labels generated by \code{\link{gt_render}}.}
  \item{index}{Index}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
