\name{[.HeatmapList}
\alias{[.HeatmapList}
\alias{Extract.HeatmapList}
\title{
Subset a HeatmapList object
}
\description{
Subset a HeatmapList object
}
\usage{
\method{[}{HeatmapList}(x, i, j)
}
\arguments{

  \item{x}{A \code{\link{HeatmapList-class}} object}
  \item{i}{row indices}
  \item{j}{column indices}

}
\details{
If the heatmap list is horizontal, \code{i} is the real row indices and \code{j} corresponds to heatmap names and single annotation names.
}
\examples{
# There is no example
NULL

}
