\name{full_comb_code}
\alias{full_comb_code}
\title{
Full set of code of combination sets
}
\description{
Full set of code of combination sets
}
\usage{
full_comb_code(n, complement = FALSE)
}
\arguments{

  \item{n}{Number of sets}
  \item{complement}{Whether include the code for complement set?}

}
\examples{
full_comb_code(2)
full_comb_code(3)
full_comb_code(4)
full_comb_code(4, TRUE)
}
