\name{get_color_mapping_param_list-HeatmapAnnotation-method}
\alias{get_color_mapping_param_list,HeatmapAnnotation-method}
\alias{get_color_mapping_param_list}
\title{
Get a list of color mapping parameters
}
\description{
Get a list of color mapping parameters
}
\usage{
\S4method{get_color_mapping_param_list}{HeatmapAnnotation}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapAnnotation-class}} object.}

}
\details{
Color mapping parameters for visible simple annotations are only returned.

This function is only for internal use.
}
\value{
A list.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
