\name{get_legend_param_list-HeatmapAnnotation-method}
\alias{get_legend_param_list,HeatmapAnnotation-method}
\alias{get_legend_param_list}
\title{
Get a List of Annotation Legend Parameters
}
\description{
Get a List of Annotation Legend Parameters
}
\usage{
\S4method{get_legend_param_list}{HeatmapAnnotation}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapAnnotation-class}} object.}

}
\details{
The annotation legend parameters for visible simple annotations are only returned.

This function is only for internal use.
}
\value{
A list.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
