\name{grid.text_box}
\alias{grid.text_box}
\title{
Draw multiple texts in a box
}
\description{
Draw multiple texts in a box
}
\usage{
grid.text_box(text, x = unit(0.5, "npc"), y = unit(0.5, "npc"), gp = gpar(), ...)
}
\arguments{

  \item{text}{A vector of texts. The value can be single words or phrases/sentenses.}
  \item{x}{X position.}
  \item{y}{Y position.}
  \item{gp}{Graphics parameters of texts.}
  \item{...}{Pass to \code{\link{text_box_grob}}.}

}
\details{
All details can be found in the help page of \code{\link{text_box_grob}}.
}
\examples{
# There is no example
NULL

}
