\name{Heatmap-class}
\docType{class}
\alias{Heatmap-class}
\alias{Heatmap}
\title{
class for single heatmap  


}
\description{
class for single heatmap  


}
\details{
The components for a heamtap is  

  \preformatted{
         +------+
         +------+
         +------+
         +------+
   +-+-+-+------+-+-+-+
   | | | |      | | | |
   +-+-+-+------+-+-+-+
         +------+
         +------+
         +------+
         +------+
  }


}
