\name{heatmap_legend_size-HeatmapList-method}
\alias{heatmap_legend_size,HeatmapList-method}
\title{
get heatmap legend size  


}
\description{
get heatmap legend size  


}
\usage{
\S4method{heatmap_legend_size}{HeatmapList}(object, side = c("right", "left", "top", "bottom"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{side}{side}

}
