\name{HeatmapAnnotation-class}
\docType{class}
\alias{HeatmapAnnotation-class}
\alias{HeatmapAnnotation}
\title{
Class for heatmap annotations  


}
\description{
Class for heatmap annotations  


}
\details{
A complex heatmap contains a list of annotation which represented as different graphics placed on rows and columns. The \code{\link{HeatmapAnnotation}} class stores information by a list of \code{\link{SingleAnnotation}} objects.  


}
\section{Methods}{
The \code{\link{HeatmapAnnotation}} class provides following methods:  

\itemize{
  \item \code{\link{initialize,HeatmapAnnotation-method}}: constructor method
  \item \code{\link{get_color_mapping_list,HeatmapAnnotation-method}}
  \item \code{\link{draw,HeatmapAnnotation-method}}
}


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
