\name{HeatmapList-class}
\docType{class}
\alias{HeatmapList-class}
\alias{HeatmapList}
\title{
class for a list of heatmaps  


}
\description{
class for a list of heatmaps  


}
\details{
  \preformatted{
         +------+
         +------+
         +------+
   +-+-+-+------+-+-+-+
   | | | |      | | | |
   +-+-+-+------+-+-+-+
         +------+
         +------+
         +------+
  }


}
