\name{height.Heatmap}
\alias{height.Heatmap}
\title{
Height of the Heatmap
}
\description{
Height of the Heatmap
}
\usage{
\method{height}{Heatmap}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapList-class}} object returned by \code{\link{draw,Heatmap-method}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
