\name{height.HeatmapAnnotation}
\alias{height.HeatmapAnnotation}
\title{
Height of the HeatmapAnnotation Object
}
\description{
Height of the HeatmapAnnotation Object
}
\usage{
\method{height}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL
}
