\name{height.HeatmapList}
\alias{height.HeatmapList}
\title{
Height of the Heatmap List
}
\description{
Height of the Heatmap List
}
\usage{
\method{height}{HeatmapList}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapList-class}} object returned by \code{\link{draw,HeatmapList-method}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
