\name{height.SingleAnnotation}
\alias{height.SingleAnnotation}
\title{
Height of the SingleAnnotation object
}
\description{
Height of the SingleAnnotation object
}
\usage{
\method{height}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
