\name{height<-.AnnotationFunction}
\alias{height<-.AnnotationFunction}
\alias{heightAssign.AnnotationFunction}
\title{
Assign the Height to the AnnotationFunction x
}
\description{
Assign the Height to the AnnotationFunction x
}
\usage{
\method{height}{AnnotationFunction}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{AnnotationFunction-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
