\name{height<-}
\alias{height<-}
\alias{heightAssign}
\title{
Generic Assignment Method for height()
}
\description{
Generic Assignment Method for height()
}
\usage{
height(x, ...) <- value
}
\arguments{

  \item{x}{An object.}
  \item{value}{The value}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL
}
