\name{height<-.SingleAnnotation}
\alias{height<-.SingleAnnotation}
\alias{heightAssign.SingleAnnotation}
\title{
Assign the Height to the SingleAnnotation x
}
\description{
Assign the Height to the SingleAnnotation x
}
\usage{
\method{height}{SingleAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
