\name{heightDetails.annotation_axis}
\alias{heightDetails.annotation_axis}
\title{
Height for annotation_axis Grob
}
\description{
Height for annotation_axis Grob
}
\usage{
\method{heightDetails}{annotation_axis}(x)
}
\arguments{

  \item{x}{The \code{annotation_axis} grob returned by \code{\link{annotation_axis_grob}}.}

}
\details{
The physical height of the grob can be get by \code{convertWidth(grobHeight(axis_grob), "mm")}.
}
\examples{
# There is no example
NULL

}
