\name{heightDetails.packed_legends}
\alias{heightDetails.packed_legends}
\title{
Grob height for packed_legends
}
\description{
Grob height for packed_legends
}
\usage{
\method{heightDetails}{packed_legends}(x)
}
\arguments{

  \item{x}{A packed_legends object.}

}
\examples{
# There is no example
NULL

}
