\name{heightDetails.textbox}
\alias{heightDetails.textbox}
\title{
Height for textbox grob
}
\description{
Height for textbox grob
}
\usage{
\method{heightDetails}{textbox}(x)
}
\arguments{

  \item{x}{The \code{textbox} grob returned by \code{\link{textbox_grob}}.}

}
\value{
A \code{\link[grid]{unit}} object.
}
\examples{
# There is no example
NULL

}
