\name{ht_size}
\alias{ht_size}
\title{
Calculate the width and height of the heatmaps
}
\description{
Calculate the width and height of the heatmaps
}
\usage{
ht_size(ht)
}
\arguments{

  \item{ht}{A \code{\link{Heatmap-class}} or \code{\link{HeatmapList-class}} object.}

}
\value{
A list of two elements: width and height.
}
\examples{
# There is no example
NULL

}
