\name{initialize-ColorMapping-method}
\alias{initialize,ColorMapping-method}
\title{
initialize  


}
\description{
initialize  


}
\usage{
\S4method{initialize}{ColorMapping}(.Object, name, colors = NULL, levels = NULL,
    col_fun = NULL, breaks = NULL)
}
\arguments{

  \item{name}{name}
  \item{colors}{colors}
  \item{levels}{levels}
  \item{col_fun}{color function}
  \item{breaks}{breaks}

}
