\name{initialize-dispatch}
\alias{initialize}
\title{
Method dispatch page for initialize
}
\description{
Method dispatch page for \code{initialize}.
}
\section{Dispatch}{
\code{initialize} can be dispatched on following classes:

\itemize{
\item \code{\link{initialize,ColorMapping-method}}, \code{ColorMapping} class method
\item \code{\link{initialize,Heatmap-method}}, \code{Heatmap} class method
}

}
