\name{Legends}
\alias{Legends}
\title{
Constructor method for Legends class
}
\description{
Constructor method for Legends class
}
\usage{
Legends(...)
}
\arguments{

  \item{...}{arguments.}

}
\details{
There is no public constructor method for the \code{\link{Legends-class}}.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
