\name{length.HeatmapAnnotation}
\alias{length.HeatmapAnnotation}
\title{
Number of Annotations
}
\description{
Number of Annotations
}
\usage{
\method{length}{HeatmapAnnotation}(x)
}
\arguments{

  \item{x}{A \code{\link{HeatmapAnnotation-class}} object.}

}
\examples{
# There is no example
NULL

}
