\name{list_components}
\alias{list_components}
\title{
List All Heatmap Components
}
\description{
List All Heatmap Components
}
\usage{
list_components(pattern = NULL)
}
\arguments{

  \item{pattern}{A regular expression.}

}
\value{
A vector of viewport names.
}
\examples{
# There is no example
NULL

}
