% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upset.R
\name{make_comb_mat_from_matrix}
\alias{make_comb_mat_from_matrix}
\title{Complement Set Size}
\usage{
make_comb_mat_from_matrix(
  x,
  mode,
  top_n_sets = Inf,
  min_set_size = -Inf,
  universal_set = NULL,
  complement_size = NULL,
  set_on_rows = TRUE
)
}
\arguments{
\item{m}{A combination matrix returned by \code{\link{make_comb_mat}}.}

\item{degree}{degree of the intersection. The value can be a vector.}
}
\value{
If there is no complement set, it returns zero.

If there is no complement set, it returns zero.

If there is no complement set, it returns zero.

A vector of sizes of the combination sets.

A vector of sizes of the combination sets.

A vector of sizes of the combination sets.

If there is no complement set, it returns zero.

If there is no complement set, it returns zero.

If there is no complement set, it returns zero.

A vector of sizes of the combination sets.

A vector of sizes of the combination sets.

A vector of sizes of the combination sets.
}
\description{
Complement Set Size
}
\examples{


# There is no example
NULL


Complement Set Size

Complement Set Size



# There is no example
NULL


Complement Set Size

Complement Set Size



# There is no example
NULL


Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)

Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)

Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)


Complement Set Size

Complement Set Size



# There is no example
NULL


Complement Set Size

Complement Set Size



# There is no example
NULL


Complement Set Size

Complement Set Size



# There is no example
NULL


Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)

Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)

Sizes of the Combination sets

Sizes of the Combination sets



set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_size(m)

}
