\name{make_layout-dispatch}
\alias{make_layout}
\title{
Method dispatch page for make_layout
}
\description{
Method dispatch page for \code{make_layout}.
}
\section{Dispatch}{
\code{make_layout} can be dispatched on following classes:

\itemize{
\item \code{\link{make_layout,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{make_layout,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
