\name{map-ColorMapping-method}
\alias{map,ColorMapping-method}
\alias{map}
\title{
Map values to colors  


}
\description{
Map values to colors  


}
\usage{
\S4method{map}{ColorMapping}(object, x)
}
\arguments{

  \item{object}{a \code{\link{ColorMapping}} object.}
  \item{x}{input values.}

}
\details{
It maps a vector of values to a vector of colors.  


}
\value{
A vector of colors.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
