\name{nobs.HeatmapAnnotation}
\alias{nobs.HeatmapAnnotation}
\title{
Number of Observations
}
\description{
Number of Observations
}
\usage{
\method{nobs}{HeatmapAnnotation}(object, ...)
}
\arguments{

  \item{object}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{other arguments.}

}
\examples{
# There is no example
NULL

}
