\name{nobs.SingleAnnotation}
\alias{nobs.SingleAnnotation}
\title{
Number of Observations
}
\description{
Number of Observations
}
\usage{
\method{nobs}{SingleAnnotation}(object, ...)
}
\arguments{

  \item{object}{The \code{\link{SingleAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\details{
It returns the \code{n} slot of the annotaton function. If it does not exist, it returns \code{NA}.
}
\examples{
# There is no example
NULL

}
