\name{normalize_comb_mat}
\alias{normalize_comb_mat}
\title{
Normalize a list of combination matrice
}
\description{
Normalize a list of combination matrice
}
\usage{
normalize_comb_mat(...)
}
\arguments{

  \item{...}{If it is a single argument, the value should be a list of combination matrices.}

}
\details{
It normalizes a list of combination matrice to make them have same number and order of sets and combination sets.

The sets (by \code{\link{set_name}}) from all combination matrice should be the same.
}
\examples{
# There is no example
NULL

}
