\name{plot.Heatmap}
\alias{plot.Heatmap}
\title{
Draw heatmap
}
\description{
Draw heatmap
}
\usage{
\method{plot}{Heatmap}(x, ...)
}
\arguments{

  \item{x}{A \code{\link{Heatmap-class}} object.}
  \item{...}{All pass to \code{\link{draw,Heatmap-method}}.}

}
\examples{
# There is no example
NULL

}
