\name{plot.HeatmapAnnotation}
\alias{plot.HeatmapAnnotation}
\title{
Draw heatmap annotations
}
\description{
Draw heatmap annotations
}
\usage{
\method{plot}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{A \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{All pass to \code{\link{draw,HeatmapList-method}}.}

}
\examples{
# There is no example
NULL

}
