\name{plot.HeatmapList}
\alias{plot.HeatmapList}
\title{
Draw heatmap
}
\description{
Draw heatmap
}
\usage{
\method{plot}{HeatmapList}(x, ...)
}
\arguments{

  \item{x}{A \code{\link{HeatmapList-class}} object.}
  \item{...}{All pass to \code{\link{draw,HeatmapList-method}}.}

}
\examples{
# There is no example
NULL

}
