\name{print.comb_mat}
\alias{print.comb_mat}
\title{
Print the comb_mat Object
}
\description{
Print the comb_mat Object
}
\usage{
\method{print}{comb_mat}(x, ...)
}
\arguments{

  \item{x}{A combination matrix returned by \code{\link{make_comb_mat}}.}
  \item{...}{Other arguments}

}
\examples{
# There is no example
NULL

}
