\name{row_anno_barplot}
\alias{row_anno_barplot}
\title{
Barplots as Row Annotation
}
\description{
Barplots as Row Annotation
}
\usage{
row_anno_barplot(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_barplot}}.}

}
\details{
A wrapper of \code{\link{anno_barplot}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_barplot}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_barplot}}.
}
\examples{
# There is no example
NULL

}
