\name{row_anno_boxplot}
\alias{row_anno_boxplot}
\title{
Boxplots as Row Annotation
}
\description{
Boxplots as Row Annotation
}
\usage{
row_anno_boxplot(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_boxplot}}.}

}
\details{
A wrapper of \code{\link{anno_boxplot}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_boxplot}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_boxplot}}.
}
\examples{
# There is no example
NULL

}
