\name{row_anno_density}
\alias{row_anno_density}
\title{
Density as Row Annotation
}
\description{
Density as Row Annotation
}
\usage{
row_anno_density(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_density}}.}

}
\details{
A wrapper of \code{\link{anno_density}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_density}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_density}}.
}
\examples{
# There is no example
NULL

}
