\name{row_anno_link}
\alias{row_anno_link}
\title{
Label Markers as Row Annotation
}
\description{
Label Markers as Row Annotation
}
\usage{
row_anno_link(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_link}}.}

}
\details{
A wrapper of \code{\link{anno_link}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_link}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_link}}.
}
\examples{
# There is no example
NULL

}
