\name{row_anno_points}
\alias{row_anno_points}
\title{
Row annotation which is represented as points
}
\description{
Row annotation which is represented as points
}
\usage{
row_anno_points(...)}
\arguments{

  \item{...}{pass to \code{\link{anno_points}}}
}
\details{
A wrapper of \code{\link{anno_points}} with pre-defined \code{which} to \code{row}.
}
