\name{row_dend-dispatch}
\alias{row_dend}
\title{
Method dispatch page for row_dend
}
\description{
Method dispatch page for \code{row_dend}.
}
\section{Dispatch}{
\code{row_dend} can be dispatched on following classes:

\itemize{
\item \code{\link{row_dend,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{row_dend,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
