\name{row_order-dispatch}
\alias{row_order}
\title{
Method dispatch page for row_order
}
\description{
Method dispatch page for \code{row_order}.
}
\section{Dispatch}{
\code{row_order} can be dispatched on following classes:

\itemize{
\item \code{\link{row_order,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{row_order,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
