\name{select_region}
\alias{select_region}
\title{
Select a region in the heatmap
}
\description{
Select a region in the heatmap
}
\usage{
select_region(mark = FALSE)
}
\arguments{

  \item{mark}{whether show the rectangular selected region}

}
\details{
Users can use mouse to click two positions on the heatmap, the function
will return the row index and column index for the selected region in the selected matrix.
}
