\name{set_component_height-Heatmap-method}
\alias{set_component_height,Heatmap-method}
\title{
Set height of each heatmap component  


}
\description{
Set height of each heatmap component  


}
\usage{
\S4method{set_component_height}{Heatmap}(object, k, v)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object.}
  \item{k}{which components, see \code{\link{Heatmap-class}}.}
  \item{v}{height of the component, a \code{\link[grid]{unit}} object.}

}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\alias{set_component_height}
