\name{show-AnnotationFunction-method}
\alias{show,AnnotationFunction-method}
\title{
Print the AnnotationFunction object
}
\description{
Print the AnnotationFunction object
}
\usage{
\S4method{show}{AnnotationFunction}(object)
}
\arguments{

  \item{object}{The \code{\link{AnnotationFunction-class}} object.}

}
\examples{
# There is no example
NULL
}
