\name{show-ColorMapping-method}
\alias{show,ColorMapping-method}
\title{
print ColorMapping object  


}
\description{
print ColorMapping object  


}
\usage{
\S4method{show}{ColorMapping}(object)
}
\arguments{

  \item{object}{a \code{\link{ColorMapping}} object.}

}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
