\name{show-dispatch}
\alias{show}
\title{
Method dispatch page for show
}
\description{
Method dispatch page for \code{show}.
}
\section{Dispatch}{
\code{show} can be dispatched on following classes:

\itemize{
\item \code{\link{show,ColorMapping-method}}, \code{ColorMapping} class method
\item \code{\link{show,HeatmapList-method}}, \code{HeatmapList} class method
\item \code{\link{show,Heatmap-method}}, \code{Heatmap} class method
}

}
