\name{show-HeatmapAnnotation-method}
\alias{show,HeatmapAnnotation-method}
\title{
Print the HeatmapAnnotation object
}
\description{
Print the HeatmapAnnotation object
}
\usage{
\S4method{show}{HeatmapAnnotation}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapAnnotation-class}} object.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
