\name{show-SingleAnnotation-method}
\alias{show,SingleAnnotation-method}
\title{
Print the SingleAnnotation object
}
\description{
Print the SingleAnnotation object
}
\usage{
\S4method{show}{SingleAnnotation}(object)
}
\arguments{

  \item{object}{a \code{\link{SingleAnnotation-class}} object.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
