\name{size.HeatmapAnnotation}
\alias{size.HeatmapAnnotation}
\title{
Size of the HeatmapAnnotation x
}
\description{
Size of the HeatmapAnnotation x
}
\usage{
\method{size}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} x.}
  \item{...}{other arguments}

}
\details{
It returns the width if it is a row annotation and the height if it is a column annotation.
}
\examples{
# There is no example
NULL

}
