\name{size.SingleAnnotation}
\alias{size.SingleAnnotation}
\title{
Size of the SingleAnnotation x
}
\description{
Size of the SingleAnnotation x
}
\usage{
\method{size}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} x.}
  \item{...}{other arguments}

}
\details{
It returns the width if it is a row annotation and the height if it is a column annotation.
}
\examples{
# There is no example
NULL
}
