\name{size<-}
\alias{size<-}
\alias{sizeAssign}
\title{
Generic Assignment Method for size()
}
\description{
Generic Assignment Method for size()
}
\usage{
size(x, ...) <- value
}
\arguments{

  \item{x}{An object.}
  \item{value}{The value}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
