\name{str.comb_mat}
\alias{str.comb_mat}
\title{
str method
}
\description{
str method
}
\usage{
\method{str}{comb_mat}(object, ...)
}
\arguments{

  \item{object}{A combination matrix returned by \code{\link{make_comb_mat}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
