\name{subset_matrix_by_row}
\alias{subset_matrix_by_row}
\title{
Subset the Matrix by Rows
}
\description{
Subset the Matrix by Rows
}
\usage{
subset_matrix_by_row(x, i)
}
\arguments{

  \item{x}{A matrix.}
  \item{i}{The row indices.}

}
\details{
Mainly used for constructing the \code{\link{AnnotationFunction-class}} object.
}
\examples{
# There is no example
NULL

}
