\name{subset_no}
\alias{subset_no}
\title{
Do not do subseting
}
\description{
Do not do subseting
}
\usage{
subset_no(x, i)
}
\arguments{

  \item{x}{A vector.}
  \item{i}{The indices.}

}
\details{
Mainly used for constructing the \code{\link{AnnotationFunction-class}} object.
}
\examples{
# There is no example
NULL

}
