\name{subset_vector}
\alias{subset_vector}
\title{
Subset the vector
}
\description{
Subset the vector
}
\usage{
subset_vector(x, i)
}
\arguments{

  \item{x}{A vector.}
  \item{i}{The indices.}

}
\details{
Mainly used for constructing the \code{\link{AnnotationFunction-class}} object.
}
\examples{
# There is no example
NULL

}
