\name{summary.Heatmap}
\alias{summary.Heatmap}
\title{
Print the Summary of a Heatmap
}
\description{
Print the Summary of a Heatmap
}
\usage{
\method{summary}{Heatmap}(object, ...)
}
\arguments{

  \item{object}{A \code{\link{Heatmap-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
