\name{summary.HeatmapList}
\alias{summary.HeatmapList}
\title{
Summary of a Heatmap List
}
\description{
Summary of a Heatmap List
}
\usage{
\method{summary}{HeatmapList}(object, ...)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
