\name{unify_mat_list}
\alias{unify_mat_list}
\title{
Unify a List of Matrix
}
\description{
Unify a List of Matrix
}
\usage{
unify_mat_list(mat_list, default = 0)
}
\arguments{

  \item{mat_list}{A list of matrix. All of them should have dimension names.}
  \item{default}{Default values for the newly added rows and columns.}

}
\details{
All matrix will be unified to have same row names and column names.
}
\value{
A list of matrix
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
