\name{width.Heatmap}
\alias{width.Heatmap}
\title{
Width of the Heatmap
}
\description{
Width of the Heatmap
}
\usage{
\method{width}{Heatmap}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapList-class}} object returned by \code{\link{draw,Heatmap-method}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
