\name{width.HeatmapAnnotation}
\alias{width.HeatmapAnnotation}
\title{
Width of the HeatmapAnnotation x
}
\description{
Width of the HeatmapAnnotation x
}
\usage{
\method{width}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
