\name{width.HeatmapList}
\alias{width.HeatmapList}
\title{
Width of the Heatmap List
}
\description{
Width of the Heatmap List
}
\usage{
\method{width}{HeatmapList}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapList-class}} object returned by \code{\link{draw,HeatmapList-method}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
