\name{width.Legends}
\alias{width.Legends}
\title{
Width of the Legends
}
\description{
Width of the Legends
}
\usage{
\method{width}{Legends}(x, ...)
}
\arguments{

  \item{x}{The \code{\link[grid]{grob}} x returned by \code{\link{Legend}} or \code{\link{packLegend}}.}
  \item{...}{other arguments}

}
\value{
The returned unit x is in \code{mm}.
}
\examples{
# There is no example
NULL

}
