\name{width.SingleAnnotation}
\alias{width.SingleAnnotation}
\title{
Width of the SingleAnnotation x
}
\description{
Width of the SingleAnnotation x
}
\usage{
\method{width}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL
}
