\name{width<-.AnnotationFunction}
\alias{width<-.AnnotationFunction}
\alias{widthAssign.AnnotationFunction}
\title{
Assign the Width to the AnnotationFunction x
}
\description{
Assign the Width to the AnnotationFunction x
}
\usage{
\method{width}{AnnotationFunction}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{AnnotationFunction-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
