\name{width<-.HeatmapAnnotation}
\alias{width<-.HeatmapAnnotation}
\alias{widthAssign.HeatmapAnnotation}
\title{
Assign the Width to the HeatmapAnnotation x
}
\description{
Assign the Width to the HeatmapAnnotation x
}
\usage{
\method{width}{HeatmapAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
