\name{width<-}
\alias{width<-}
\alias{widthAssign}
\title{
Generic Assignment Method for width()
}
\description{
Generic Assignment Method for width()
}
\usage{
width(x, ...) <- value
}
\arguments{

  \item{x}{An object.}
  \item{value}{The value}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL
}
