\name{width<-.SingleAnnotation}
\alias{width<-.SingleAnnotation}
\title{
Assign the Width to the SingleAnnotation x
}
\description{
Assign the Width to the SingleAnnotation x
}
\usage{
\method{width}{SingleAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL
}
