\name{widthDetails.annotation_axis}
\alias{widthDetails.annotation_axis}
\title{
Width for annotation_axis Grob
}
\description{
Width for annotation_axis Grob
}
\usage{
\method{widthDetails}{annotation_axis}(x)
}
\arguments{

  \item{x}{The \code{annotation_axis} grob returned by \code{\link{annotation_axis_grob}}.}

}
\details{
The physical width of the grob can be get by \code{convertWidth(grobWidth(axis_grob), "mm")}.
}
\examples{
# There is no example
NULL

}
