\name{widthDetails.packed_legends}
\alias{widthDetails.packed_legends}
\title{
Grob width for packed_legends
}
\description{
Grob width for packed_legends
}
\usage{
\method{widthDetails}{packed_legends}(x)
}
\arguments{

  \item{x}{A packed_legends object.}

}
\examples{
# There is no example
NULL

}
