\name{widthDetails.text_box}
\alias{widthDetails.text_box}
\title{
Width for text_box grob
}
\description{
Width for text_box grob
}
\usage{
\method{widthDetails}{text_box}(x)
}
\arguments{

  \item{x}{The \code{text_box} grob returned by \code{\link{text_box_grob}}.}

}
\value{
A \code{\link[grid]{unit}} object.
}
\examples{
# There is no example
NULL

}
