\name{widthDetails.textbox}
\alias{widthDetails.textbox}
\title{
Width for textbox grob
}
\description{
Width for textbox grob
}
\usage{
\method{widthDetails}{textbox}(x)
}
\arguments{

  \item{x}{The \code{textbox} grob returned by \code{\link{textbox_grob}}.}

}
\value{
A \code{\link[grid]{unit}} object.
}
\examples{
# There is no example
NULL

}
