\name{+.Heatmap}
\alias{+.Heatmap}
\title{
combine two heatmaps as a heatmap list  


}
\description{
combine two heatmaps as a heatmap list  


}
\usage{
\method{+}{Heatmap}(ht1, ht2)
}
\arguments{

  \item{ht1}{a \code{\link{Heatmap}} object}
  \item{ht2}{a \code{\link{Heatmap}} object}

}
\value{
a \code{\link{HeatmapList}} object  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
