\name{+.Heatmap}
\alias{+.Heatmap}
\title{
Add two heatmaps as a heatmap list  


}
\description{
Add two heatmaps as a heatmap list  


}
\usage{
\method{+}{Heatmap}(x, y)
}
\arguments{

  \item{x}{a \code{\link{Heatmap}} object.}
  \item{y}{a \code{\link{Heatmap}} object or a \code{\link{HeatmapList}} object.}

}
\details{
It is only a shortcut function. It actually calls \code{\link{add_heatmap,Heatmap-method}}.  


}
\value{
a \code{\link{HeatmapList}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
