\name{+.HeatmapList}
\alias{+.HeatmapList}
\title{
add heatmaps to the list  


}
\description{
add heatmaps to the list  


}
\usage{
\method{+}{HeatmapList}(ht1, ht2)
}
\arguments{

  \item{ht1}{a \code{\link{HeatmapList}} object}
  \item{ht2}{a \code{\link{Heatmap}} object or a \code{\link{HeatmapList}} object}

}
\value{
a \code{\link{HeatmapList}} object  


}
