\name{+.HeatmapList}
\alias{+.HeatmapList}
\title{
Add heatmaps to the list  


}
\description{
Add heatmaps to the list  


}
\usage{
\method{+}{HeatmapList}(ht1, ht2)
}
\arguments{

  \item{ht1}{a \code{\link{HeatmapList}} object.}
  \item{ht2}{a \code{\link{Heatmap}} object or a \code{\link{HeatmapList}} object.}

}
\value{
A \code{\link{HeatmapList}} object  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
