\name{add_heatmap-Heatmap-method}
\alias{add_heatmap,Heatmap-method}
\title{
Add two heatmaps as a heatmap list  


}
\description{
Add two heatmaps as a heatmap list  


}
\usage{
\S4method{add_heatmap}{Heatmap}(object, ht)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object.}
  \item{ht}{a \code{\link{Heatmap}} object.}

}
\details{
There is a shortcut function \code{\link{+.Heatmap}}.  


}
\value{
A \code{\link{HeatmapList}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
