\name{add_heatmap-HeatmapList-method}
\alias{add_heatmap,HeatmapList-method}
\title{
Add heatmaps and row annotations to the heatmap list  


}
\description{
Add heatmaps and row annotations to the heatmap list  


}
\usage{
\S4method{add_heatmap}{HeatmapList}(object, x)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object.}
  \item{x}{a \code{\link{Heatmap}} object or a \code{\link{HeatmapAnnotation}} object or a \code{\link{HeatmapList}} object.}

}
\details{
There is a shortcut function \code{+.HeatmapList}.  


}
\value{
A \code{\link{HeatmapList}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
