\name{add_heatmap-HeatmapList-method}
\alias{add_heatmap,HeatmapList-method}
\title{
add heatmaps to the heatmap list  


}
\description{
add heatmaps to the heatmap list  


}
\usage{
\S4method{add_heatmap}{HeatmapList}(object, ht)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{ht}{a \code{\link{Heatmap}} object or a \code{\link{HeatmapList}} object}

}
\value{
a \code{\link{HeatmapList}} object  


}
