\name{AdditiveUnit-class}
\docType{class}
\alias{AdditiveUnit-class}
\alias{AdditiveUnit}
\title{
An internal class  


}
\description{
An internal class  


}
\details{
This class is a super class for \code{\link{Heatmap}}, \code{\link{HeatmapList}} and \code{\link{HeatmapAnnotation}} classes. And it is only for \code{+} generic method.  


}
