\name{AdditiveUnit-class}
\docType{class}
\alias{AdditiveUnit-class}
\title{
An internal class  


}
\description{
An internal class  


}
\details{
This class is a super class for \code{\link{Heatmap-class}}, \code{\link{HeatmapList-class}} and \code{\link{HeatmapAnnotation-class}} classes. It is only designed for \code{+} generic method.  


}
\examples{
# no example
NULL
}
