% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditiveUnit-class.R
\docType{class}
\name{AdditiveUnit-class}
\alias{AdditiveUnit-class}
\title{Class for Concatenating Heatmaps and Annotations}
\description{
Class for Concatenating Heatmaps and Annotations
}
\details{
This class is a super class for \code{\link{Heatmap-class}},
\code{\link{HeatmapList-class}} and \code{\link{HeatmapAnnotation-class}}
classes. It is only designed for \code{+} generic method and the \code{\%v\%v}
method so that above three classes can be appended to each other.
}
\examples{

# There is no example
NULL


}
