\name{AdditiveUnit}
\alias{AdditiveUnit}
\title{
Constructor method for AdditiveUnit class
}
\description{
Constructor method for AdditiveUnit class
}
\usage{
AdditiveUnit(...)
}
\arguments{

  \item{...}{black hole arguments.}

}
\details{
This method is not used in the package.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this function
NULL
}
