% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditiveUnit-class.R
\name{AdditiveUnit}
\alias{AdditiveUnit}
\title{Constructor Method for AdditiveUnit Class}
\usage{
AdditiveUnit(...)
}
\arguments{
\item{...}{Black hole arguments.}
}
\value{
No value is returned.
}
\description{
Constructor Method for AdditiveUnit Class
}
\details{
This method is not used in the package.
}
\examples{


# There is no example
NULL


}
\author{
Zuguang Gu \href{mailto:z.gu@dkfz.dez.gu@dkfz.de}
}
