\name{adjust_heatmap_list-HeatmapList-method}
\alias{adjust_heatmap_list,HeatmapList-method}
\title{
Adjust heatmap list
}
\description{
Adjust heatmap list
}
\usage{
\S4method{adjust_heatmap_list}{HeatmapList}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object}

}
\alias{adjust_heatmap_list}
\examples{
# There is no example
NULL
}
