\name{adjust_heatmap_list-HeatmapList-method}
\alias{adjust_heatmap_list,HeatmapList-method}
\title{
Adjust heatmap list
}
\description{
Adjust heatmap list
}
\usage{
\S4method{adjust_heatmap_list}{HeatmapList}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object}

}
\details{
This function adjust heatmap components for all heatmaps to align them nicely.

This function is only for internal use.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\alias{adjust_heatmap_list}
\examples{
# There is no example
NULL
}
