\name{adjust_heatmap_list-HeatmapList-method}
\alias{adjust_heatmap_list,HeatmapList-method}
\alias{adjust_heatmap_list}
\title{
Adjust Heatmap List
}
\description{
Adjust Heatmap List
}
\usage{
\S4method{adjust_heatmap_list}{HeatmapList}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}

}
\details{
This function adjusts settings in all other heatmaps according to the main heatmap.
It also adjust the size of heatmap annotations to make them aligned nicely.

This function is only for internal use.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
