\name{anno_link}
\alias{anno_link}
\title{
Link Annotation
}
\description{
Link Annotation
}
\usage{
anno_link(...)
}
\arguments{

  \item{...}{Pass to \code{\link{anno_zoom}}.}

}
\details{
This function is the same as \code{\link{anno_zoom}}. It links subsets of rows or columns to a list of graphic regions.
}
\examples{
# There is no example
NULL

}
