% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFunction-function.R
\name{anno_link}
\alias{anno_link}
\title{Link Annotation}
\usage{
anno_link(...)
}
\arguments{
\item{...}{Pass to \code{\link{anno_zoom}}.}
}
\description{
Link Annotation
}
\details{
This function is the same as \code{\link{anno_zoom}}. It links subsets of
rows or columns to a list of graphic regions.
}
\examples{


# There is no example
NULL



}
