\name{anno_oncoprint_barplot}
\alias{anno_oncoprint_barplot}
\title{
Barplot annotation for oncoPrint
}
\description{
Barplot annotation for oncoPrint
}
\usage{
anno_oncoprint_barplot(type = all_type, which = c("column", "row"),
    width = NULL, height = NULL, border = FALSE, ...)
}
\arguments{

  \item{type}{-type}
  \item{which}{-which}
  \item{width}{-width}
  \item{height}{-height}
  \item{border}{-border}
  \item{...}{-...}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
