\name{anno_oncoprint_barplot}
\alias{anno_oncoprint_barplot}
\title{
<<<<<<< HEAD
Column barplot annotation for oncoPrint
}
\description{
Column barplot annotation for oncoPrint
=======
Column barplot annotaiton for oncoPrint
}
\description{
Column barplot annotaiton for oncoPrint
>>>>>>> bioc/master
}
\usage{
anno_oncoprint_barplot()
}
\details{
This function is only used for column annotation
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
