\name{annotation_legend_size-HeatmapList-method}
\alias{annotation_legend_size,HeatmapList-method}
\alias{annotation_legend_size}
\title{
Size of the annotation legend viewport  


}
\description{
Size of the annotation legend viewport  


}
\usage{
\S4method{annotation_legend_size}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}

}
\details{
This function is only for internal use.  


}
\value{
A \code{\link[grid]{unit}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
