\name{column_dend-Heatmap-method}
\alias{column_dend,Heatmap-method}
\title{
Get column dendrograms from a heatmap
}
\description{
Get column dendrograms from a heatmap
}
\usage{
\S4method{column_dend}{Heatmap}(object)
}
\arguments{

  \item{object}{a \code{\link{Heatmap-class}} object}

}
\value{
A dendrogram object
}
\examples{
# There is no example
NULL

}
