\name{column_order-Heatmap-method}
\alias{column_order,Heatmap-method}
\title{
Get column order from a heatmap list
}
\description{
Get column order from a heatmap list
}
\usage{
\S4method{column_order}{Heatmap}(object)
}
\arguments{

  \item{object}{a \code{\link{Heatmap-class}} object}

}
\value{
A vector containing column orders
}
\examples{
# There is no example
NULL

}
