\name{column_order-HeatmapList-method}
\alias{column_order,HeatmapList-method}
\title{
Get column order from a heatmap list
}
\description{
Get column order from a heatmap list
}
\usage{
\S4method{column_order}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object}

}
\value{
A list contains column orders which correspond every matrix
}
\examples{
# There is no example
NULL

}
