% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapAnnotation-class.R
\name{columnAnnotation}
\alias{columnAnnotation}
\title{Construct Column Annotations}
\usage{
columnAnnotation(...)
}
\arguments{
\item{...}{Pass to \code{\link{HeatmapAnnotation}}.}
}
\value{
A \code{\link{HeatmapAnnotation-class}} object.
}
\description{
Construct Column Annotations
}
\details{
The function is identical to

\preformatted{ HeatmapAnnotation(..., which = "column") }
}
\examples{


# There is no example
NULL



}
\author{
Zuguang Gu \href{mailto:z.gu@dkfz.dez.gu@dkfz.de}
}
