% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upset.R
\name{comb_degree}
\alias{comb_degree}
\title{Degrees of the Combination sets}
\usage{
comb_degree(m)
}
\arguments{
\item{m}{A combination matrix returned by \code{\link{make_comb_mat}}.}
}
\value{
A vector of degrees of the combination sets.
}
\description{
Degrees of the Combination sets
}
\details{
The degree for a combination set is the number of sets that are selected.
}
\examples{


set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_degree(m)


}
