% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_translation.R
\name{compare_heatmap}
\alias{compare_heatmap}
\title{Compare heatmaps between stats::heatmap() and ComplexHeatmap::heatmap()}
\usage{
compare_heatmap(...)
}
\arguments{
\item{...}{The same set of arguments passed to \code{stats::heatmap} and
\code{ComplexHeatmap::heatmap}.}
}
\description{
Compare heatmaps between stats::heatmap() and ComplexHeatmap::heatmap()
}
\details{
The function plots two heatmaps, one by \code{stats::heatmap} and one by
\code{ComplexHeatmap::heatmap}. Users can see the difference between the two
implementations.
}
\examples{


mat = matrix(rnorm(100), 10)
compare_heatmap(mat)


}
