\name{component_height-dispatch}
\alias{component_height}
\title{
Method dispatch page for component_height
}
\description{
Method dispatch page for \code{component_height}.
}
\section{Dispatch}{
\code{component_height} can be dispatched on following classes:

\itemize{
\item \code{\link{component_height,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{component_height,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
