\name{component_height-Heatmap-method}
\alias{component_height,Heatmap-method}
\title{
Height of each heatmap component
}
\description{
Height of each heatmap component
}
\usage{
\S4method{component_height}{Heatmap}(object, k = 1:9)
}
\arguments{

  \item{object}{a \code{\link{Heatmap-class}} object.}
  \item{k}{which component in the heatmap, see \code{\link{Heatmap-class}}.}

}
\details{
This function is only for internal use.
}
\value{
A \code{\link[grid]{unit}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this internal method

}
