\name{component_height-HeatmapList-method}
\alias{component_height,HeatmapList-method}
\title{
Height of Heatmap List Components
}
\description{
Height of Heatmap List Components
}
\usage{
\S4method{component_height}{HeatmapList}(object, k = HEATMAP_LIST_LAYOUT_COLUMN_COMPONENT)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}
  \item{k}{Which component in the heatmap list. Values are in \code{ComplexHeatmap:::HEATMAP_LIST_LAYOUT_COLUMN_COMPONENT}.}

}
\value{
A \code{\link[grid]{unit}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
