\name{component_height-HeatmapList-method}
\alias{component_height,HeatmapList-method}
\title{
Height of each heatmap list component
}
\description{
Height of each heatmap list component
}
\usage{
\S4method{component_height}{HeatmapList}(object, k = 1:7)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object.}
  \item{k}{which component in the heatmap list, see \code{\link{HeatmapList-class}}.}

}
\value{
A \code{\link[grid]{unit}} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this internal method

}
