\name{component_width-Heatmap-method}
\alias{component_width,Heatmap-method}
\title{
Width of each heatmap component  


}
\description{
Width of each heatmap component  


}
\usage{
\S4method{component_width}{Heatmap}(object, k = 1:7)
}
\arguments{

  \item{object}{a \code{\link{Heatmap-class}} object.}
  \item{k}{which components, see \code{\link{Heatmap-class}}.}

}
\section{Detials}{
This function is only for internal use.  


}
\value{
A \code{\link[grid]{unit}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
