\name{component_width-HeatmapList-method}
\alias{component_width,HeatmapList-method}
\title{
Width of each heatmap list component
}
\description{
Width of each heatmap list component
}
\usage{
\S4method{component_width}{HeatmapList}(object, k = HEATMAP_LIST_LAYOUT_ROW_COMPONENT)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object.}
  \item{k}{which component in the heatmap list, see \code{\link{HeatmapList-class}}.}

}
\details{
This function is only for internal use.
}
\value{
A \code{\link[grid]{unit}} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
