\name{component_width-HeatmapList-method}
\alias{component_width,HeatmapList-method}
\title{
Width of Heatmap List Components
}
\description{
Width of Heatmap List Components
}
\usage{
\S4method{component_width}{HeatmapList}(object, k = HEATMAP_LIST_LAYOUT_ROW_COMPONENT)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}
  \item{k}{Which component in the heatmap list. Values are in \code{ComplexHeatmap:::HEATMAP_LIST_LAYOUT_ROW_COMPONENT}.}

}
\details{
This function is only for internal use.
}
\value{
A \code{\link[grid]{unit}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
