% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_column_dend}
\alias{decorate_column_dend}
\title{Decorate Heatmap Column Dendrograms}
\usage{
decorate_column_dend(..., envir = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Pass to \code{\link{decorate_dend}}.}

\item{envir}{Where to look for variables inside \code{code}.}
}
\value{
The function returns no value.
}
\description{
Decorate Heatmap Column Dendrograms
}
\details{
This is a wrapper function which pre-defined \code{which} argument in
\code{\link{decorate_dend}}.
}
\examples{


# There is no example
NULL



}
\author{
Zuguang Gu \href{mailto:z.gu@dkfz.dez.gu@dkfz.de}
}
