\name{decorate_column_title}
\alias{decorate_column_title}
\title{
Decorate heatmap column title
}
\description{
Decorate heatmap column title
}
\usage{
decorate_column_title(..., envir = new.env(parent = parent.frame()))
}
\arguments{

  \item{...}{pass to \code{\link{decorate_title}}}
  \item{envir}{where to look for variables inside \code{code}}

}
\details{
This is a helper function which pre-defined \code{which} argument in \code{\link{decorate_title}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# No example for this function
NULL
}
