\name{decorate_row_dend}
\alias{decorate_row_dend}
\title{
Decorate heatmap dendrogram on rows
}
\description{
Decorate heatmap dendrogram on rows
}
\usage{
decorate_row_dend(..., envir = new.env(parent = parent.frame()))
}
\arguments{

  \item{...}{pass to \code{\link{decorate_dend}}}
  \item{envir}{where to look for variables inside \code{code}}

}
\details{
This is a helper function which pre-defined \code{which} argument in \code{\link{decorate_dend}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# No example for this function
NULL
}
