\name{decorate_row_names}
\alias{decorate_row_names}
\title{
Decorate Heatmap Row Names
}
\description{
Decorate Heatmap Row Names
}
\usage{
decorate_row_names(..., envir = new.env(parent = parent.frame()))
}
\arguments{

  \item{...}{Pass to \code{\link{decorate_dimnames}}.}
  \item{envir}{wWhere to look for variables inside \code{code}.}

}
\details{
This is a helper function which pre-defined \code{which} argument in \code{\link{decorate_dimnames}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
