% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_row_names}
\alias{decorate_row_names}
\title{Decorate Heatmap Row Names}
\usage{
decorate_row_names(..., envir = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Pass to \code{\link{decorate_dimnames}}.}

\item{envir}{wWhere to look for variables inside \code{code}.}
}
\value{
The function returns no value.
}
\description{
Decorate Heatmap Row Names
}
\details{
This is a helper function which pre-defined \code{which} argument in
\code{\link{decorate_dimnames}}.
}
\examples{


# There is no example
NULL



}
\author{
Zuguang Gu \href{mailto:z.gu@dkfz.dez.gu@dkfz.de}
}
