\name{decorate_row_title}
\alias{decorate_row_title}
\title{
Decorate Heatmap Row Titles
}
\description{
Decorate Heatmap Row Titles
}
\usage{
decorate_row_title(..., envir = new.env(parent = parent.frame()))
}
\arguments{

  \item{...}{Pass to \code{\link{decorate_title}}.}
  \item{envir}{Where to look for variables inside \code{code}.}

}
\details{
This is a helper function which pre-defined \code{which} argument in \code{\link{decorate_title}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
