\name{decorate_row_title}
\alias{decorate_row_title}
\title{
Decorate heatmap row title
}
\description{
Decorate heatmap row title
}
\usage{
decorate_row_title(...)
}
\arguments{

  \item{...}{pass to \code{\link{decorate_title}}}

}
\details{
This is a helper function which pre-defined \code{which} argument in \code{\link{decorate_title}}.
}
\value{
The function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# No example for this function
NULL
}
