% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_row_title}
\alias{decorate_row_title}
\title{Decorate Heatmap Row Titles}
\usage{
decorate_row_title(..., envir = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Pass to \code{\link{decorate_title}}.}

\item{envir}{Where to look for variables inside \code{code}.}
}
\value{
The function returns no value.
}
\description{
Decorate Heatmap Row Titles
}
\details{
This is a helper function which pre-defined \code{which} argument in
\code{\link{decorate_title}}.
}
\examples{


# There is no example
NULL



}
\author{
Zuguang Gu \href{mailto:z.gu@dkfz.dez.gu@dkfz.de}
}
