% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoPrint.R
\name{default_get_type}
\alias{default_get_type}
\title{Default get_type for oncoPrint()}
\usage{
default_get_type(x)
}
\arguments{
\item{x}{A strings which encode multiple altertations.}
}
\description{
Default get_type for oncoPrint()
}
\details{
It recognizes following separators: \code{;:,|}.
}
\examples{


# There is no example
NULL



}
