\name{draw-dispatch}
\alias{draw}
\title{
Method dispatch page for draw
}
\description{
Method dispatch page for \code{draw}.
}
\section{Dispatch}{
\code{draw} can be dispatched on following classes:

\itemize{
\item \code{\link{draw,HeatmapList-method}}, \code{HeatmapList} class method
\item \code{\link{draw,Heatmap-method}}, \code{Heatmap} class method
}

}
