\name{draw-dispatch}
\alias{draw}
\title{
Method dispatch page for draw
}
\description{
Method dispatch page for \code{draw}.
}
\section{Dispatch}{
\code{draw} can be dispatched on following classes:

\itemize{
\item \code{\link{draw,HeatmapAnnotation-method}}, \code{\link{HeatmapAnnotation-class}} class method
\item \code{\link{draw,SingleAnnotation-method}}, \code{\link{SingleAnnotation-class}} class method
\item \code{\link{draw,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{draw,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
