\name{draw-Heatmap-method}
\alias{draw,Heatmap-method}
\title{
draw heatmap  


}
\description{
draw heatmap  


}
\usage{
\S4method{draw}{Heatmap}(object, internal = FALSE, ...)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object}
  \item{internal}{internal use}
  \item{...}{pass to \code{\link{HeatmapList}}}

}
