\name{draw-HeatmapAnnotation-method}
\alias{draw,HeatmapAnnotation-method}
\title{
Draw the heatmap annotations  


}
\description{
Draw the heatmap annotations  


}
\usage{
\S4method{draw}{HeatmapAnnotation}(object, index, ...)
}
\arguments{

  \item{object}{a \code{\link{HeatmapAnnotation}} object}
  \item{index}{the index of rows or columns}
  \item{...}{pass to \code{\link[grid]{viewport}} which contains all annotations.}

}
\details{
A viewport is created.  


}
\value{
No value is returned.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
