\name{draw_annotation-Heatmap-method}
\alias{draw_annotation,Heatmap-method}
\alias{draw_annotation}
\title{
Draw column annotations  


}
\description{
Draw column annotations  


}
\usage{
\S4method{draw_annotation}{Heatmap}(object, which = c("top", "bottom"))
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object.}
  \item{which}{on top or bottom}

}
\details{
A viewport is created which contains column annotations.  

This function is only for internal use.  


}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
