\name{draw_annotation-Heatmap-method}
\alias{draw_annotation,Heatmap-method}
\title{
Draw column annotations  


}
\description{
Draw column annotations  


}
\usage{
\S4method{draw_annotation}{Heatmap}(object, gp = object@gp_list$colnames_annotation_gp)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object.}
  \item{gp}{graphic parameters for drawing rectangles.}

}
\details{
A viewport is created which contains column annotations.  


}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\alias{draw_annotation}
