\name{draw_annotation_legend-HeatmapList-method}
\alias{draw_annotation_legend,HeatmapList-method}
\title{
draw annotation legend  


}
\description{
draw annotation legend  


}
\usage{
\S4method{draw_annotation_legend}{HeatmapList}(object, side = c("right", "left", "top", "bottom"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{side}{side}

}
\alias{draw_annotation_legend}
