\name{draw_annotation_legend-HeatmapList-method}
\alias{draw_annotation_legend,HeatmapList-method}
\alias{draw_annotation_legend}
\title{
Draw legends for all column annotations  


}
\description{
Draw legends for all column annotations  


}
\usage{
\S4method{draw_annotation_legend}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}

}
\details{
A viewport is created which contains annotation legends.  

This function is only for internal use.  


}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
