\name{draw_heatmap_body-Heatmap-method}
\alias{draw_heatmap_body,Heatmap-method}
\title{
plot the heatmap body  


}
\description{
plot the heatmap body  


}
\usage{
\S4method{draw_heatmap_body}{Heatmap}(object, gp = object@gp_list$rect_gp)
}
\arguments{

  \item{object}{a \code{\link{Heatmap}} object}
  \item{gp}{graphic parameters for drawing rectangles}

}
\details{
the heatmap body 100% covers the viewport  


}
\value{
this function returns no value  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
