\name{draw_heatmap_legend-HeatmapList-method}
\alias{draw_heatmap_legend,HeatmapList-method}
\title{
Draw legends for all heatmaps  


}
\description{
Draw legends for all heatmaps  


}
\usage{
\S4method{draw_heatmap_legend}{HeatmapList}(object, side = c("right", "left", "top", "bottom"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{side}{side}

}
\details{
A viewport is created which contains heatmap legends.  


}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\alias{draw_heatmap_legend}
