\name{draw_heatmap_list-HeatmapList-method}
\alias{draw_heatmap_list,HeatmapList-method}
\alias{draw_heatmap_list}
\title{
Draw the list of heatmaps  


}
\description{
Draw the list of heatmaps  


}
\usage{
\S4method{draw_heatmap_list}{HeatmapList}(object, hgap = unit(2, "mm"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}
  \item{hgap}{gap}

}
\details{
A viewport is created which contains heatmaps.  


}
\value{
This function returns no value.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
