\name{draw_heatmap_list-HeatmapList-method}
\alias{draw_heatmap_list,HeatmapList-method}
\alias{draw_heatmap_list}
\title{
Draw the List of Heatmaps
}
\description{
Draw the List of Heatmaps
}
\usage{
\S4method{draw_heatmap_list}{HeatmapList}(object)
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}

}
\details{
It only draws the list of heatmaps without legends and titles.

This function is only for internal use.
}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
