\name{draw_heatmap_list-HeatmapList-method}
\alias{draw_heatmap_list,HeatmapList-method}
\alias{draw_heatmap_list}
\title{
Draw the list of heatmaps
}
\description{
Draw the list of heatmaps
}
\usage{
\S4method{draw_heatmap_list}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object}

}
\details{
A viewport is created which contains heatmaps.

This function is only for internal use.
}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this internal method
NULL

}
