\name{draw_title-dispatch}
\alias{draw_title}
\title{
Method dispatch page for draw_title
}
\description{
Method dispatch page for \code{draw_title}.
}
\section{Dispatch}{
\code{draw_title} can be dispatched on following classes:

\itemize{
\item \code{\link{draw_title,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{draw_title,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
