\name{draw_title-HeatmapList-method}
\alias{draw_title,HeatmapList-method}
\title{
Draw Heatmap List Title
}
\description{
Draw Heatmap List Title
}
\usage{
\S4method{draw_title}{HeatmapList}(object,
    which = c("column", "row"))
}
\arguments{

  \item{object}{A \code{\link{HeatmapList-class}} object.}
  \item{which}{Is it a row title or a column title.}

}
\details{
A viewport is created which contains heatmap list title.

This function is only for internal use.
}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
