\name{draw_title-HeatmapList-method}
\alias{draw_title,HeatmapList-method}
\title{
Draw heatmap list title
}
\description{
Draw heatmap list title
}
\usage{
\S4method{draw_title}{HeatmapList}(object,
    which = c("column", "row"))
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object}
  \item{which}{dendrogram on the row or on the column of the heatmap}

}
\details{
A viewport is created which contains heatmap list title.

This function is only for internal use.
}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this internal method
NULL

}
