\name{get_color_mapping_list-HeatmapAnnotation-method}
\alias{get_color_mapping_list,HeatmapAnnotation-method}
\alias{get_color_mapping_list}
\title{
Get a list of color mapping objects  


}
\description{
Get a list of color mapping objects  


}
\usage{
\S4method{get_color_mapping_list}{HeatmapAnnotation}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapAnnotation}} object.}

}
\details{
Color mapping for simple annotations are returned.  


}
\value{
A list of \code{\link{ColorMapping}} list.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
