\name{heatmap_legend_size-HeatmapList-method}
\alias{heatmap_legend_size,HeatmapList-method}
\alias{heatmap_legend_size}
\title{
Size of the heatmap legend viewprot  


}
\description{
Size of the heatmap legend viewprot  


}
\usage{
\S4method{heatmap_legend_size}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}

}
\details{
This function is only for internal use.  


}
\value{
A \code{\link[grid]{unit}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
