\name{HeatmapAnnotation-class}
\docType{class}
\alias{HeatmapAnnotation-class}
\title{
Class for heatmap annotations  


}
\description{
Class for heatmap annotations  


}
\details{
A complex heatmap contains a list of annotations which represente as different graphics placed on rows and columns. The \code{\link{HeatmapAnnotation-class}} is a category of single annotations by a list of \code{\link{SingleAnnotation-class}} objects with same number of rows or columns.  


}
\section{Methods}{
The \code{\link{HeatmapAnnotation-class}} provides following methods:  

\itemize{
  \item \code{\link{HeatmapAnnotation}}: constructor method
  \item \code{\link{draw,HeatmapAnnotation-method}}
}


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
