\name{HeatmapAnnotation-class}
\docType{class}
\alias{HeatmapAnnotation-class}
\title{
Class for Heatmap Annotations
}
\description{
Class for Heatmap Annotations
}
\details{
A complex heatmap contains a list of annotations which are represented as graphics
placed on rows and columns. The \code{\link{HeatmapAnnotation-class}} contains a list of single annotations which are
represented as a list of \code{\link{SingleAnnotation-class}} objects.
}
\section{Methods}{
The \code{\link{HeatmapAnnotation-class}} provides following methods:

\itemize{
  \item \code{\link{HeatmapAnnotation}}: constructor method.
  \item \code{\link{draw,HeatmapAnnotation-method}}: draw the annotations.
}}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
