\name{HeatmapList-class}
\docType{class}
\alias{HeatmapList-class}
\title{
Class for a list of heatmaps
}
\description{
Class for a list of heatmaps
}
\details{
A heatmap list is defined as a list of heatmaps and annotations.
}
\section{Methods}{
The \code{\link{HeatmapList-class}} provides following methods:

\itemize{
  \item \code{\link{draw,HeatmapList-method}}: draw the list of heatmaps and row annotations.
  \item \code{\link{add_heatmap,HeatmapList-method}}: add heatmaps to the list of heatmaps.
  \item \code{\link{row_order,HeatmapList-method}}: get order of rows
  \item \code{\link{column_order,HeatmapList-method}}: get order of columns
  \item \code{\link{row_dend,HeatmapList-method}}: get row dendrograms
  \item \code{\link{column_dend,HeatmapList-method}}: get column dendrograms
}}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
