\name{HeatmapList}
\alias{HeatmapList}
\title{
Constructor method for HeatmapList class  


}
\description{
Constructor method for HeatmapList class  


}
\usage{
HeatmapList(...)
}
\arguments{

  \item{...}{arguments}

}
\details{
There is no public constructor method for the \code{\link{HeatmapList-class}}.  


}
\section{Detailes}{
There is no public constructor method for the \code{\link{HeatmapList-class}}.
}
