\name{height.HeatmapAnnotation}
\alias{height.HeatmapAnnotation}
\title{
Height of the HeatmapAnnotation x
}
\description{
Height of the HeatmapAnnotation x
}
\usage{
\method{height}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
