\name{height.Legends}
\alias{height.Legends}
\title{
Height of the Legends
}
\description{
Height of the Legends
}
\usage{
\method{height}{Legends}(x, ...)
}
\arguments{

  \item{x}{The \code{\link[grid]{grob}} x returned by \code{\link{Legend}} or \code{\link{packLegend}}.}
  \item{...}{other arguments}

}
\value{
The returned unit x is in \code{mm}.
}
\examples{
# There is no example
NULL

}
