\name{height.SingleAnnotation}
\alias{height.SingleAnnotation}
\title{
Height of the SingleAnnotation x
}
\description{
Height of the SingleAnnotation x
}
\usage{
\method{height}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL
}
