\name{heightAssign.AnnotationFunction}
\alias{heightAssign.AnnotationFunction}
\alias{height<-.AnnotationFunction}
\title{
Assign the Height to the AnnotationFunction Object
}
\description{
Assign the Height to the AnnotationFunction Object
}
\usage{
\method{height}{AnnotationFunction}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{AnnotationFunction-class}} object.}
  \item{value}{A \code{\link[grid]{unit}} object.}
  \item{...}{Other arguments.}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
