\name{height<-.HeatmapAnnotation}
\alias{height<-.HeatmapAnnotation}
\alias{heightAssign.HeatmapAnnotation}
\title{
Assign the Height to the HeatmapAnnotation x
}
\description{
Assign the Height to the HeatmapAnnotation x
}
\usage{
\method{height}{HeatmapAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} x.}
  \item{value}{A \code{\link[grid]{unit}} x.}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
