\name{heightAssign.HeatmapAnnotation}
\alias{heightAssign.HeatmapAnnotation}
\alias{height<-.HeatmapAnnotation}
\title{
Assign the Height to the HeatmapAnnotation Object
}
\description{
Assign the Height to the HeatmapAnnotation Object
}
\usage{
\method{height}{HeatmapAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{value}{A \code{\link[grid]{unit}} object.}
  \item{...}{Other arguments.}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
