\name{height<-.SingleAnnotation}
\alias{height<-.SingleAnnotation}
\alias{heightAssign.SingleAnnotation}
\title{
Assign the Height to the SingleAnnotation Object
}
\description{
Assign the Height to the SingleAnnotation Object
}
\usage{
\method{height}{SingleAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} object.}
  \item{value}{A \code{\link[grid]{unit}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL
}
