\name{initialize-ColorMapping-method}
\alias{initialize,ColorMapping-method}
\title{
Constructor methods for ColorMapping class  


}
\description{
Constructor methods for ColorMapping class  


}
\usage{
\S4method{initialize}{ColorMapping}(.Object, name, colors = NULL, levels = NULL,
    col_fun = NULL, breaks = NULL)
}
\arguments{

  \item{.Object}{private object}
  \item{name}{name for this color mapping. It is used for drawing the title of the legend.}
  \item{colors}{discrete colors.}
  \item{levels}{levels that correspond to \code{colors}. If \code{colors} is name indexed,  \code{levels} can be ignored.}
  \item{col_fun}{color mapping function that maps continuous values to colors.}
  \item{breaks}{breaks for the continuous color mapping. If \code{col_fun} is generated by \code{\link[circlize]{colorRamp2}}, \code{breaks} can be ignored.}

}
\details{
\code{colors} and \code{levels} are used for discrete color mapping, \code{col_fun} and  \code{breaks} are used for continuous color mapping.  


}
\value{
A \code{\link{ColorMapping}} object.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
