\name{initialize-dispatch}
\alias{initialize}
\title{
Method dispatch page for initialize
}
\description{
Method dispatch page for \code{initialize}.
}
\section{Dispatch}{
\code{initialize} can be dispatched on following classes:

\itemize{
\item \code{\link{initialize,Heatmap-method}}, \code{\link{Heatmap}} class method
\item \code{\link{initialize,SingleAnnotation-method}}, \code{\link{SingleAnnotation}} class method
\item \code{\link{initialize,HeatmapAnnotation-method}}, \code{\link{HeatmapAnnotation}} class method
\item \code{\link{initialize,ColorMapping-method}}, \code{\link{ColorMapping}} class method
}

}
