\name{make_column_cluster-Heatmap-method}
\alias{make_column_cluster,Heatmap-method}
\alias{make_column_cluster}
\title{
make column cluster  


}
\description{
make column cluster  


}
\usage{
\S4method{make_column_cluster}{Heatmap}(object, order = "hclust", distance = object@column_hclust_param$distance,
    method = object@column_hclust_param$method)
}
