\name{map-ColorMapping-method}
\alias{map,ColorMapping-method}
\title{
map values to colors  


}
\description{
map values to colors  


}
\usage{
\S4method{map}{ColorMapping}(object, x)
}
\arguments{

  \item{object}{a \code{\link{ColorMapping}} object}
  \item{x}{input values}

}
\details{
It maps a vector of values to a vector of colors  


}
\value{
a vector of colors  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\alias{map}
