\name{nobs.HeatmapAnnotation}
\alias{nobs.HeatmapAnnotation}
\title{
Number of Observations
}
\description{
Number of Observations
}
\usage{
\method{nobs}{HeatmapAnnotation}(object, ...)
}
\arguments{

  \item{object}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{other arguments.}

}
\value{
If there is no \code{nobs} information for any of its \code{\link{SingleAnnotation-class}} object,
it returns \code{NA}.
}
\examples{
# There is no example
NULL

}
