\name{packLegend}
\alias{packLegend}
\title{
Pack legends
}
\description{
Pack legends
}
\usage{
packLegend(..., gap = unit(4, "mm"), direction = c("vertical", "horizontal"))
}
\arguments{

  \item{...}{objects returned by \code{\link{Legend}}}
  \item{gap}{gap between two legends. The value is a \code{\link[grid]{unit}} object}
  \item{direction}{how to arrange legends}

}
\value{
A \code{\link[grid]{grob}} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
