\name{row_anno_barplot}
\alias{row_anno_barplot}
\title{
Row annotation which is represented as barplots
}
\description{
Row annotation which is represented as barplots
}
\usage{
row_anno_barplot(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_barplot}}}

}
\details{
A wrapper of \code{\link{anno_barplot}} with pre-defined \code{which} to \code{row}.
}
\value{
See help page of \code{\link{anno_barplot}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
