\name{row_anno_density}
\alias{row_anno_density}
\title{
Row annotation which is represented as density plot
}
\description{
Row annotation which is represented as density plot
}
\usage{
row_anno_density(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_density}}}

}
\details{
A wrapper of \code{\link{anno_density}} with pre-defined \code{which} to \code{row}.
}
\value{
See help page of \code{\link{anno_density}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
