\name{row_anno_histogram}
\alias{row_anno_histogram}
\title{
Histograms as Row Annotation
}
\description{
Histograms as Row Annotation
}
\usage{
row_anno_histogram(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_histogram}}.}

}
\details{
A wrapper of \code{\link{anno_histogram}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_histogram}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_histogram}}.
}
\examples{
# There is no example
NULL

}
