\name{row_anno_histogram}
\alias{row_anno_histogram}
\title{
Row annotation which is represented as histogram
}
\description{
Row annotation which is represented as histogram
}
\usage{
row_anno_histogram(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_histogram}}}

}
\details{
A wrapper of \code{\link{anno_histogram}} with pre-defined \code{which} to \code{row}.
}
\value{
See help page of \code{\link{anno_histogram}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
