\name{row_anno_link}
\alias{row_anno_link}
\title{
Column annotation which is represented as links
}
\description{
Column annotation which is represented as links
}
\usage{
row_anno_link(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_link}}}

}
\details{
A wrapper of \code{\link{anno_link}} with pre-defined \code{which} to \code{row}.
}
\value{
See help page of \code{\link{anno_link}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
