\name{row_anno_points}
\alias{row_anno_points}
\title{
Points as Row Annotation
}
\description{
Points as Row Annotation
}
\usage{
row_anno_points(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_points}}.}

}
\details{
A wrapper of \code{\link{anno_points}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_points}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_points}}.
}
\examples{
# There is no example
NULL

}
