\name{row_anno_text}
\alias{row_anno_text}
\title{
Text as Row Annotation
}
\description{
Text as Row Annotation
}
\usage{
row_anno_text(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_text}}.}

}
\details{
A wrapper of \code{\link{anno_text}} with pre-defined \code{which} to \code{row}.

You can directly use \code{\link{anno_text}} for row annotation if you call it in \code{\link{rowAnnotation}}.
}
\value{
See help page of \code{\link{anno_text}}.
}
\examples{
# There is no example
NULL

}
