\name{row_anno_text}
\alias{row_anno_text}
\title{
Row annotation which is represented as text
}
\description{
Row annotation which is represented as text
}
\usage{
row_anno_text(...)
}
\arguments{

  \item{...}{pass to \code{\link{anno_text}}}

}
\details{
A wrapper of \code{\link{anno_text}} with pre-defined \code{which} to \code{row}.
}
\value{
See help page of \code{\link{anno_text}}
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
