\name{row_dend-Heatmap-method}
\alias{row_dend,Heatmap-method}
\title{
Get row dendrograms from a heatmap
}
\description{
Get row dendrograms from a heatmap
}
\usage{
\S4method{row_dend}{Heatmap}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList}} object}

}
\value{
A list of dendrograms for which each dendrogram corresponds to a row slice
}
\examples{
# There is no example
NULL

}
