\name{row_dend-HeatmapList-method}
\alias{row_dend,HeatmapList-method}
\title{
Get row dendrograms from a heatmap list
}
\description{
Get row dendrograms from a heatmap list
}
\usage{
\S4method{row_dend}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object}

}
\value{
A list of dendrograms for which each dendrogram corresponds to a row slice
}
\examples{
# There is no example
NULL

}
