\name{row_order-HeatmapList-method}
\alias{row_order,HeatmapList-method}
\title{
Get row order from a heatmap list
}
\description{
Get row order from a heatmap list
}
\usage{
\S4method{row_order}{HeatmapList}(object)
}
\arguments{

  \item{object}{a \code{\link{HeatmapList-class}} object}

}
\value{
A list contains row orders which correspond to the original matrix
}
\examples{
# There is no example
NULL

}
