\name{rowAnnotation}
\alias{rowAnnotation}
\title{
Construct Row Annotations
}
\description{
Construct Row Annotations
}
\usage{
rowAnnotation(...)
}
\arguments{

  \item{...}{Pass to \code{\link{HeatmapAnnotation}}.}

}
\details{
The function is identical to

  \preformatted{
    HeatmapAnnotation(..., which = "row")  }
}
\value{
A \code{\link{HeatmapAnnotation-class}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
