\name{set_component_height-Heatmap-method}
\alias{set_component_height,Heatmap-method}
\alias{set_component_height}
\title{
Set height of each heatmap component
}
\description{
Set height of each heatmap component
}
\usage{
\S4method{set_component_height}{Heatmap}(object, k, v)
}
\arguments{

  \item{object}{a \code{\link{Heatmap-class}} object.}
  \item{k}{which components, see \code{\link{Heatmap-class}}.}
  \item{v}{height of the component, a \code{\link[grid]{unit}} object.}

}
\details{
This function is only for internal use.
}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# no example for this internal method
NULL

}
