\name{show-ColorMapping-method}
\alias{show,ColorMapping-method}
\title{
Print ColorMapping object
}
\description{
Print ColorMapping object
}
\usage{
\S4method{show}{ColorMapping}(object)
}
\arguments{

  \item{object}{a \code{\link{ColorMapping-class}} object.}

}
\value{
This function returns no value.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
