\name{show-dispatch}
\alias{show}
\title{
Method dispatch page for show
}
\description{
Method dispatch page for \code{show}.
}
\section{Dispatch}{
\code{show} can be dispatched on following classes:

\itemize{
\item \code{\link{show,ColorMapping-method}}, \code{\link{ColorMapping-class}} class method
\item \code{\link{show,HeatmapAnnotation-method}}, \code{\link{HeatmapAnnotation-class}} class method
\item \code{\link{show,SingleAnnotation-method}}, \code{\link{SingleAnnotation-class}} class method
\item \code{\link{show,HeatmapList-method}}, \code{\link{HeatmapList-class}} class method
\item \code{\link{show,Heatmap-method}}, \code{\link{Heatmap-class}} class method
}
}
\examples{
# no example
NULL


}
