\name{size.HeatmapAnnotation}
\alias{size.HeatmapAnnotation}
\title{
Size of the HeatmapAnnotation Object
}
\description{
Size of the HeatmapAnnotation Object
}
\usage{
\method{size}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\details{
It returns the width if it is a row annotation and the height if it is a column annotation.

Internally used.
}
\examples{
# There is no example
NULL

}
