\name{size.SingleAnnotation}
\alias{size.SingleAnnotation}
\title{
Size of the SingleAnnotation Object
}
\description{
Size of the SingleAnnotation Object
}
\usage{
\method{size}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\details{
It returns the width if it is a row annotation and the height if it is a column annotation.

Internally used.
}
\examples{
# There is no example
NULL

}
