\name{width.HeatmapAnnotation}
\alias{width.HeatmapAnnotation}
\title{
Width of the HeatmapAnnotation Object
}
\description{
Width of the HeatmapAnnotation Object
}
\usage{
\method{width}{HeatmapAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
