\name{width.SingleAnnotation}
\alias{width.SingleAnnotation}
\title{
Width of the SingleAnnotation Object
}
\description{
Width of the SingleAnnotation Object
}
\usage{
\method{width}{SingleAnnotation}(x, ...)
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
