\name{widthAssign.AnnotationFunction}
\alias{widthAssign.AnnotationFunction}
\alias{width<-.AnnotationFunction}
\title{
Assign the Width to the AnnotationFunction Object
}
\description{
Assign the Width to the AnnotationFunction Object
}
\usage{
\method{width}{AnnotationFunction}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{AnnotationFunction-class}} object.}
  \item{...}{Other arguments.}
  \item{value}{A \code{\link[grid]{unit}} object.}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
