\name{widthAssign.HeatmapAnnotation}
\alias{widthAssign.HeatmapAnnotation}
\alias{width<-.HeatmapAnnotation}
\title{
Assign the Width to the HeatmapAnnotation Object
}
\description{
Assign the Width to the HeatmapAnnotation Object
}
\usage{
\method{width}{HeatmapAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{HeatmapAnnotation-class}} object.}
  \item{value}{A \code{\link[grid]{unit}} object.}
  \item{...}{Other arguments.}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
