\name{widthAssign.SingleAnnotation}
\alias{widthAssign.SingleAnnotation}
\alias{width<-.SingleAnnotation}
\title{
Assign the Width to the SingleAnnotation Object
}
\description{
Assign the Width to the SingleAnnotation Object
}
\usage{
\method{width}{SingleAnnotation}(x, ...) <- value
}
\arguments{

  \item{x}{The \code{\link{SingleAnnotation-class}} object.}
  \item{value}{A \code{\link[grid]{unit}} object.}
  \item{...}{Other arguments.}

}
\details{
Internally used.
}
\examples{
# There is no example
NULL

}
